atualizações

+++ routes/web.php
@@ -11,9 +11,9 @@ Route::group(['middleware' => 'web'], function () {
 
     Route::get('/', [App\Http\Controllers\HomeController::class, 'index']);
 });
-Route::get('/users', [UsuariosController::class, 'index']);
-Route::get('/users/new', [UsuariosController::class, 'createPage']);
-Route::post('/users/create', [UsuariosController::class, 'store']);
-Route::get('/users/edit/{id}', [UsuariosController::class, 'edit']);
-Route::put('/users/update/{id}', [UsuariosController::class, 'update']);
-Route::delete('/users/delete/{id}', [UsuariosController::class, 'destroy']);
+Route::get('/users', [UsuariosController::class, 'index'])->middleware('auth');
+Route::get('/users/new', [UsuariosController::class, 'createPage'])->middleware('auth');
+Route::post('/users/create', [UsuariosController::class, 'store'])->middleware('auth');
+Route::get('/users/edit/{id}', [UsuariosController::class, 'edit'])->middleware('auth');
+Route::put('/users/update/{id}', [UsuariosController::class, 'update'])->middleware('auth');
+Route::delete('/users/delete/{id}', [UsuariosController::class, 'destroy'])->middleware('auth');
